GVT-474 Add payfort merchant 2.0 option
