+++ lib/active_merchant/billing/gateways/payfort/payfort_codes.rb
@@ -1,0 +1,161 @@
+module ActiveMerchant #:nodoc:
+  module Billing #:nodoc:
+    module PayfortCodes #:nodoc:
+      ERROR_CODES = {
+        '00' => 'Invalid Request',
+        '01' => 'Order Stored',
+        '02' => 'Authorization Success',
+        '03' => 'Authorization Failed',
+        '04' => 'Capture Success',
+        '05' => 'Capture failed',
+        '06' => 'Refund Success',
+        '07' => 'Refund Failed',
+        '08' => 'Authorization Voided Successfully',
+        '09' => 'Authorization Void Failed',
+        '10' => 'Incomplete',
+        '11' => 'Check status Failed',
+        '12' => 'Check status success',
+        '13' => 'Purchase Failure',
+        '14' => 'Purchase Success',
+        '15' => 'Uncertain Transaction',
+        '17' => 'Tokenization failed',
+        '18' => 'Tokenization success',
+        '19' => 'Transaction pending',
+        '20' => 'On hold',
+        '21' => 'SDK token creation failure',
+        '22' => 'SDK token creation success',
+        '23' => 'Failed to process Digital Wallet service',
+        '24' => 'Digital wallet order processed successfully',
+        '27' => 'Check card balance failed',
+        '28' => 'Check card balance success.',
+        '29' => 'Redemption failed',
+        '30' => 'Redemption success',
+        '31' => 'Reverse Redemption transaction failed',
+        '32' => '	Reverse Redemption transaction success.',
+        '40' => 'Transaction In review',
+        '42' => 'Currency conversion success',
+        '43' => 'Currency conversion failed',
+        '46' => 'Bill creation success',
+        '47' => 'Bill creation failed',
+        '48' => 'Generating invoice payment link success',
+        '49' => 'Generating invoice payment link failed',
+        '50' => 'Batch file upload successfully.',
+        '51' => 'Upload batch file failed',
+        '52' => '	Token created successfully',
+        '53' => 'Token creation failed',
+        '54' => 'Get Tokens Success',
+        '55' => 'Get Tokens Failed',
+        '56' => 'Reporting Request Success',
+        '57' => 'Reporting Request Failed',
+        '58' => 'Token updated successfully',
+        '59' => 'Token updated failed',
+        '62' => 'Get Installment Plans Successfully',
+        '63' => 'Get Installment plans Failed',
+        '70' => 'Get batch results successfully',
+        '71' => 'Get batch results failed',
+        '72' => 'Batch processing success',
+        '73' => 'Batch processing failed',
+        '74' => 'Bank transfer failed',
+        '75' => 'Bank transfer successfully',
+        '76' => 'Batch validation successfully',
+        '77' => 'Batch validation failede'
+      }.freeze
+
+      SUCCESS_CODES = %w(01 02 04 06 08 12 14 18 22).freeze
+      FAILURE_CODES = %w(00 03 05 07 09 11 13 21).freeze
+
+      MESSAGE_CODES = {
+        '000' => 'Success',
+        '001' => 'Missing parameter',
+        '002' => 'Invalid parameter format',
+        '003' => 'Payment option is not available for this merchant\’s account',
+        '004' => 'Invalid command',
+        '005' => 'Invalid amount',
+        '006' => 'Technical problem',
+        '007' => 'Duplicate order number',
+        '008' => 'Signature mismatch',
+        '009' => 'Invalid merchant identifier',
+        '010' => 'Invalid access code',
+        '011' => 'Order not saved',
+        '012' => 'Card expired',
+        '013' => 'Invalid currency',
+        '014' => 'Inactive payment option',
+        '015' => 'Inactive merchant account',
+        '016' => 'Invalid card number',
+        '017' => 'Operation not allowed by the acquirer',
+        '018' => 'Operation not allowed by processor',
+        '019' => 'Inactive acquirer',
+        '020' => 'Processor is inactive',
+        '021' => 'Payment option deactivated by acquirer',
+        '022' => 'Payment option deactivated by processor',
+        '023' => 'Currency not accepted by acquirer',
+        '024' => 'Currency not accepted by processor',
+        '025' => 'Processor integration settings are missing',
+        '026' => 'Acquirer integration settings are missing',
+        '027' => 'Invalid extra parameters',
+        '028' => 'Missing operations settings.
+                  Contact PAYFORT operations support',
+        '029' => 'Insufficient funds',
+        '030' => 'Authentication failed',
+        '031' => 'Invalid issuer',
+        '032' => 'Invalid parameter length',
+        '033' => 'Parameter value not allowed',
+        '034' => 'Operation not allowed',
+        '035' => 'Order created successfully',
+        '036' => 'Order not found',
+        '038' => 'Tokenization service inactive',
+        '040' => 'Invalid transaction source as it does not match the
+                  Origin URL or the Origin IP',
+        '042' => 'Operation amount exceeds the authorized amount',
+        '043' => 'Inactive Operation',
+        '044' => 'Token name does not exist',
+        '045' => 'Merchant does not have the token service
+                  and yet "token_name" was sent',
+        '046' => 'Channel is not configured for the selected payment option',
+        '048' => 'Operation amount exceeds the captured amount',
+        '051' => 'Acquirer bank is facing technical difficulties,
+                  please try again later',
+        '052' => 'Invalid OLP',
+        '053' => 'Merchant is not found in OLP Engine DB',
+        '054' => 'SADAD is facing technical difficulties,
+                  please try again later',
+        '055' => 'OLP ID Alias is not valid. Please contact your bank',
+        '056' => 'OLP ID Alias does not exist.
+                  Please enter a valid OLP ID Alias',
+        '057' => 'Transaction amount exceeds the daily transaction limit',
+        '058' => 'Transaction amount exceeds the allowed limit per transaction',
+        '059' => 'Merchant Name and SADAD Merchant ID do not match',
+        '060' => 'The entered OLP password is incorrect.
+                  Please provide a valid password',
+        '061' => 'Failed to create token',
+        '062' => 'Token has been created',
+        '063' => 'Token has been updated',
+        '064' => '3-D Secure check requested',
+        '065' => 'Transaction waiting for payer\'s action',
+        '066' => 'Merchant reference already exists',
+        '068' => 'SDK service is inactive',
+        '070' => 'device_id mismatch',
+        '071' => 'Failed to initiate connection',
+        '072' => 'Transaction has been cancelled by the consumer',
+        '073' => 'Invalid request format',
+        '074' => 'Transaction failed',
+        '075' => 'Transaction failed',
+        '076' => 'Transaction not found in OLP',
+        '077' => 'Error transaction code not found',
+        '662' => 'Operation not allowed.
+                  The specified order is not confirmed yet',
+        '666' => 'Transaction declined',
+        '773' => 'Transaction closed',
+        '777' => 'The transaction has been processed,
+                  but failed to receive confirmation',
+        '778' => 'Session timed-out',
+        '779' => 'Transformation error',
+        '780' => 'Transaction number transformation error',
+        '781' => 'Message or response code transformation error',
+        '783' => 'Installments service inactive',
+        '785' => 'Transaction blocked by fraud checker',
+        '787' => 'Failed to authenticate the user'
+      }.freeze
+    end
+  end
+end
